+++ lib/command/setting.rb
@@ -37,8 +37,8 @@ module Command
     narou setting --list
     narou setting convert.no-open=true
     narou setting convert.no-epub=   # 右辺に何も書かないとその設定を削除できる
-    narou setting convert.copy_to=C:/dropbox/mobi
-    narou s convert.copy_to="C:\\Documents and Settings\\user\\epub"
+    narou setting convert.copy-to=C:/dropbox/mobi
+    narou s convert.copy-to="C:\\Documents and Settings\\user\\epub"
 
   Options:
       EOS
@@ -209,16 +209,20 @@ module Command
                                          " " * 6 + "※注意：KDP用のMOBIはstripしないでください"],
         "convert.no-zip" => [:boolean, "i文庫用のzipファイル作成を無効にするか"],
         "convert.no-open" => [:boolean, "変換時に保存フォルダを開かないようにするか"],
-        "convert.copy_to" => [:directory, "変換したらこのフォルダにコピーする\n" +
+        "convert.copy-to" => [:directory, "変換したらこのフォルダにコピーする\n" +
                                           " " * 6 + "※注意：存在しないフォルダだとエラーになる"],
+        "convert.copy_to" => [:directory, "copy-toの昔の書き方(非推奨)", INVISIBLE],
         "convert.inspect" => [:boolean, "常に変換時に調査結果を表示するか"],
+        "convert.multi-device" => [:string, "複数の端末用に同時に変換する。deviceよりも優先される。端末名をカンマ区切りで入力。ただのEPUBを出力したい場合はepubを指定"],
         "download.interval" => [:float, "各話DL時に指定した秒数待機する。デフォルト0"],
         "download.wait-steps" => [:integer, "この値で指定した話数ごとにウェイトを入れる\n" +
                                        " " * 6 + "※注意：11以上を設定してもなろうの場合は10話ごとにウェイトが入ります"],
         "download.use-subdirectory" => [:boolean, "小説を一定数ごとにサブフォルダへ分けて保存するか\n" +
                                                   " " * 6 + "※注意：小説を大量に同一フォルダに保存するとパフォーマンスが劣化する回避策"],
-        "send.without-freeze" => [:boolean, "`全話'送信時に凍結された小説は対象外に"],
+        "send.without-freeze" => [:boolean, "一括送信時に凍結された小説は対象外に"],
+        "send.backup-bookmark" => [:boolean, "一括送信時に栞データを自動でバックアップするか(KindlePW系用)"],
         "update.strong" => [:boolean, "更新漏れが無い様に改稿日の分は必ずDLするか"],
+        "update.logging" => [:boolean, "更新時のログを保存するかどうか"],
         "device" => [:string, "変換、送信対象の端末(sendの--help参照)"],
         "multiple-delimiter" => [:string, "--multiple指定時の区切り文字"],
       },