+++ lib/device/ibooks.rb
@@ -15,15 +15,11 @@ module Device::Ibooks
 
   RELATED_VARIABLES = {
     "force.enable_half_indent_bracket" => false,
-    "force.enable_add_date_to_title" => false,    # タイトルを変えてもiBooksに反映されないため
   }
 
   def hook_change_settings(&original_func)
     @@__already_exec_change_settings ||= false
     return if @@__already_exec_change_settings
-    force_change_settings_function({
-      "force.enable_half_indent_bracket" => false,
-    })
     @@__ibooks_container_dir = File.expand_path(IBOOKS_CONTAINER_DIR)
     unless File.exist?(@@__ibooks_container_dir)
       error "iBooksの管理フォルダが見つかりませんでした。" \