+++ lib/logger.rb
@@ -18,9 +18,18 @@ end
 module Narou end unless defined?(Narou)
 
 module Narou::LoggerModule
+  attr_accessor :capturing
+
   def initialize
     super
     @is_silent = false
+    @capturing = false
+  end
+
+  def copy_instance
+    self.class.new.tap { |obj|
+      obj.silent = silent
+    }
   end
 
   def silent=(enable)
@@ -55,18 +64,26 @@ module Narou::LoggerModule
   # 標準出力($stdout)のバッファリング＋取得
   #
   # キャプチャー用途なので標準エラーはキャプチャーしない
+  # quiet :: 標準出力に出力をしないかどうか
   # ansicolor_strip :: エスケープシーケンスを除去するか
   #
-  def capture(ansicolor_strip = true, &block)
+  def capture(options = {}, &block)
+    options = {
+      quiet: true, ansicolor_strip: true
+    }.merge(options)
     raise "#capture block given" unless block
     temp_stream = $stdout
-    $stdout = self.class.new
-    $stdout.silence do
+    $stdout = (self == $stdout ? copy_instance : self)
+    $stdout.capturing = true
+    if options[:quiet]
+      $stdout.silence { block.call }
+    else
       block.call
     end
+    $stdout.capturing = false
     buffer = $stdout.string
     $stdout = temp_stream
-    ansicolor_strip ? strip_color(buffer) : buffer
+    options[:ansicolor_strip] ? strip_color(buffer) : buffer
   end
 
   def strip_color(str)